pyproject.toml
[project]
name = "stock-intelligence-platform"
version = "0.1.0"
description = "Stock prediction platform using time-series and news sentiment"
authors = [{ name = "Sai Kartik" }]
requires-python = ">=3.10"

dependencies = [
    "pandas",
    "numpy",
    "pyyaml",
    "loguru"
]

[tool.setuptools]
package-dir = {"" = "src"}
